Extract attachment macros and method delegators into module
